requirements.txt
python --version 3.7.7
django
pillows